fix a segfault when processing truncate in cached plans (#184)
CStoreProcessUtility() removes its relations from the list of relations of
truncate statements before passing it to standard_ProcessUtility(). When
the truncate is run from a pl/pgsql function, its plan can be cached.
Since the filtered list of relations created by CStoreProcessUtility() is
freed after the execution of the truncate, it corrupts the cached plan.
To avoid this, just restore the list of relations of the node when done.